% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{appendTable}
\alias{appendTable}
\title{Append data to table.}
\usage{
appendTable(db, name, value, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{value}{Values data.frame}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbAppendTable]{DBI::dbAppendTable()}}}
}
\value{
Number of appended rows invisibly
}
\description{
Append data to table.
}
