% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{clearResult}
\alias{clearResult}
\title{Clear query or statement result.}
\usage{
clearResult(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}}
}
\value{
Invisible self
}
\description{
Clear query or statement result.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
output <- rocker::fetch(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
