% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{columnInfo}
\alias{columnInfo}
\title{Information on query result columns.}
\usage{
columnInfo(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbColumnInfo]{DBI::dbColumnInfo()}}}
}
\value{
Information table
}
\description{
Information on query result columns.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
rocker::columnInfo(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
