% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{commit}
\alias{commit}
\title{Commit transaction.}
\usage{
commit(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:transactions]{DBI::dbCommit()}}}
}
\value{
Invisible self
}
\description{
Commit transaction.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::begin(db)
rocker::sendStatement(db, "DELETE FROM mtcars WHERE gear = 3;")
rocker::clearResult(db)
rocker::commit(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
