% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{connect}
\alias{connect}
\title{Establish database connection using stored parameters.}
\usage{
connect(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
Invisible self
}
\description{
Establish database connection using stored parameters.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
