% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{createTable}
\alias{createTable}
\title{Create empty formatted table.}
\usage{
createTable(db, name, fields, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{fields}{Template data.frame}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbCreateTable]{DBI::dbCreateTable()}}}
}
\value{
Invisible self
}
\description{
Create empty formatted table.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::createTable(db, "mtcars", mtcars)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
