% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect database.}
\usage{
disconnect(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}}
}
\value{
Invisible self
}
\description{
Disconnect database.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
