% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{existsTable}
\alias{existsTable}
\title{Check if table exists.}
\usage{
existsTable(db, name, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}}}
}
\value{
TRUE or FALSE
}
\description{
Check if table exists.
}
