% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{fetch}
\alias{fetch}
\title{Fetch SQL query result from database.}
\usage{
fetch(db, n = -1, ...)
}
\arguments{
\item{db}{rocker object}

\item{n}{Number of record to be fetched}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbFetch]{DBI::dbFetch()}}}
}
\value{
Records
}
\description{
Fetch SQL query result from database.
}
