% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getInfoCon}
\alias{getInfoCon}
\title{Information on connection object.}
\usage{
getInfoCon(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetInfo]{DBI::dbGetInfo()}}}
}
\value{
Information list
}
\description{
Information on connection object.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::getInfoCon(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
