% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getInfoDrv}
\alias{getInfoDrv}
\title{Information on driver object.}
\usage{
getInfoDrv(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetInfo]{DBI::dbGetInfo()}}}
}
\value{
Information list
}
\description{
Information on driver object.
}
