% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{getInfoRes}
\alias{getInfoRes}
\title{Information on result object.}
\usage{
getInfoRes(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetInfo]{DBI::dbGetInfo()}}}
}
\value{
Information list
}
\description{
Information on result object.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
rocker::getInfoRes(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
