% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getRowCount}
\alias{getRowCount}
\title{Information on number of retrieved rows.}
\usage{
getRowCount(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetRowCount]{DBI::dbGetRowCount()}}}
}
\value{
Number of retrieved rows
}
\description{
Information on number of retrieved rows.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
output <- rocker::fetch(db)
rocker::getRowCount(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
