% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getRowsAffected}
\alias{getRowsAffected}
\title{Information on number of affected rows.}
\usage{
getRowsAffected(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetRowsAffected]{DBI::dbGetRowsAffected()}}}
}
\value{
Number of affected rows
}
\description{
Information on number of affected rows.
}
