% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getStatement}
\alias{getStatement}
\title{Information on sent statement.}
\usage{
getStatement(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetStatement]{DBI::dbGetStatement()}}}
}
\value{
Statement text
}
\description{
Information on sent statement.
}
