% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{isOpenedCon}
\alias{isOpenedCon}
\title{Check if an earlier opened connection is still open.}
\usage{
isOpenedCon(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Not used yet}
}
\value{
TRUE of FALSE
}
\description{
Check if an earlier opened connection is still open.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::isOpenedCon(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
