% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{isValidCon}
\alias{isValidCon}
\title{Check connection object.}
\usage{
isValidCon(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbIsValid]{DBI::dbIsValid()}}}
}
\value{
TRUE of FALSE
}
\description{
Check connection object.
}
