% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{isValidRes}
\alias{isValidRes}
\title{Check result object.}
\usage{
isValidRes(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbIsValid]{DBI::dbIsValid()}}}
}
\value{
TRUE of FALSE
}
\description{
Check result object.
}
