% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{listFields}
\alias{listFields}
\title{List table column names.}
\usage{
listFields(db, name, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbListFields]{DBI::dbListFields()}}}
}
\value{
Column names
}
\description{
List table column names.
}
