% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{listObjects}
\alias{listObjects}
\title{List database objects.}
\usage{
listObjects(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbListObjects]{DBI::dbListObjects()}}}
}
\value{
List of objects
}
\description{
List database objects.
}
