% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{listTables}
\alias{listTables}
\title{List database tables.}
\usage{
listTables(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbListTables]{DBI::dbListTables()}}}
}
\value{
List of objects
}
\description{
List database tables.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::listTables(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
