% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{newDB}
\alias{newDB}
\title{newDB}
\usage{
newDB(...)
}
\arguments{
\item{...}{Not used yet}
}
\value{
New instance of rocker class
}
\description{
Function generates a new \link{R6} database handling interface with \link{DBI} backend.
For more information, see \link[rocker:rocker]{rocker} class description.
}
\examples{
db <- rocker::newDB()
}
