% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{readTable}
\alias{readTable}
\title{Read table.}
\usage{
readTable(db, name, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbReadTable]{DBI::dbReadTable()}}}
}
\value{
Table
}
\description{
Read table.
}
