% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{setupSQLite}
\alias{setupSQLite}
\title{Setup database driver and define connection parameters for SQLite using \link[RSQLite:SQLite]{RSQLite} package.
Wrapper for setupDriver() function.}
\usage{
setupSQLite(db, dbname = ":memory:", protect = c("password", "user"), ...)
}
\arguments{
\item{db}{rocker object}

\item{dbname}{Database name}

\item{protect}{Parameters to be hidden}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
Invisible self
}
\description{
Setup database driver and define connection parameters for SQLite using \link[RSQLite:SQLite]{RSQLite} package.
Wrapper for setupDriver() function.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(
  db,
  dbname = ":memory:"
)
rocker::unloadDriver(db)
}
