% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{unloadDriver}
\alias{unloadDriver}
\title{Reset database driver and connection parameters.}
\usage{
unloadDriver(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, suitable parameters passed to \code{\link[DBI:dbDriver]{DBI::dbUnloadDriver()}}}
}
\value{
Invisible self
}
\description{
Reset database driver and connection parameters.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::unloadDriver(db)
}
