% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{validateCon}
\alias{validateCon}
\title{Check if an earlier opened connection is still open.}
\usage{
validateCon(db, statement = NULL, ...)
}
\arguments{
\item{db}{rocker object}

\item{statement}{Optional SQL statement. If not set default validateQuery will be used.}

\item{...}{Not used yet}
}
\value{
TRUE of FALSE
}
\description{
Check if an earlier opened connection is still open.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::validateCon(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
