% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{writeTable}
\alias{writeTable}
\title{Write data to table.}
\usage{
writeTable(db, name, value, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{value}{Values data.frame}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbWriteTable]{DBI::dbWriteTable()}}}
}
\value{
Invisible self
}
\description{
Write data to table.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
