% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{begin}
\alias{begin}
\title{Begin transaction.}
\usage{
begin(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:transactions]{DBI::dbBegin()}}}
}
\value{
Invisible self
}
\description{
Begin transaction.
}
