% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{canConnect}
\alias{canConnect}
\title{Test connection parameters.}
\usage{
canConnect(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, suitable parameters passed to \code{\link[DBI:dbCanConnect]{DBI::dbCanConnect()}}}
}
\value{
TRUE or FALSE
}
\description{
Test connection parameters.
}
