% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{columnInfo}
\alias{columnInfo}
\title{Information on query result columns.}
\usage{
columnInfo(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbColumnInfo]{DBI::dbColumnInfo()}}}
}
\value{
Information table
}
\description{
Information on query result columns.
}
