% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{commit}
\alias{commit}
\title{Commit transaction.}
\usage{
commit(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:transactions]{DBI::dbCommit()}}}
}
\value{
Invisible self
}
\description{
Commit transaction.
}
