% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getInfoRes}
\alias{getInfoRes}
\title{Information on result object.}
\usage{
getInfoRes(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetInfo]{DBI::dbGetInfo()}}}
}
\value{
Information list
}
\description{
Information on result object.
}
