% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{getRowCount}
\alias{getRowCount}
\title{Information on number of retrieved rows.}
\usage{
getRowCount(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetRowCount]{DBI::dbGetRowCount()}}}
}
\value{
Number of retrieved rows
}
\description{
Information on number of retrieved rows.
}
