% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{getRowsAffected}
\alias{getRowsAffected}
\title{Information on number of affected rows.}
\usage{
getRowsAffected(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetRowsAffected]{DBI::dbGetRowsAffected()}}}
}
\value{
Number of affected rows
}
\description{
Information on number of affected rows.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendStatement(db, "DELETE FROM mtcars WHERE gear = 3;")
rocker::getRowsAffected(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
