% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{getStatement}
\alias{getStatement}
\title{Information on sent statement.}
\usage{
getStatement(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbGetStatement]{DBI::dbGetStatement()}}}
}
\value{
Statement text
}
\description{
Information on sent statement.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
rocker::getStatement(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
