% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{hasCompleted}
\alias{hasCompleted}
\title{Information whether all records were retrieved.}
\usage{
hasCompleted(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbHasCompleted]{DBI::dbHasCompleted()}}}
}
\value{
TRUE or FALSE
}
\description{
Information whether all records were retrieved.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
output <- rocker::fetch(db, 5)
rocker::hasCompleted(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
