% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker-S3-functions.R
\name{isValidRes}
\alias{isValidRes}
\title{Check result object.}
\usage{
isValidRes(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbIsValid]{DBI::dbIsValid()}}}
}
\value{
TRUE of FALSE
}
\description{
Check result object.
}
\examples{
db <- rocker::newDB()
rocker::setupSQLite(db)
rocker::connect(db)
rocker::writeTable(db, "mtcars", mtcars)
rocker::sendQuery(db, "SELECT * FROM mtcars;")
rocker::isValidRes(db)
rocker::clearResult(db)
rocker::disconnect(db)
rocker::unloadDriver(db)
}
