% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-public.R
\name{newDB}
\alias{newDB}
\title{newDB}
\usage{
newDB(verbose = TRUE, id = NULL, ...)
}
\arguments{
\item{verbose}{TRUE or FALSE. Switch text output on / off.}

\item{id}{Optional object ID/name}

\item{...}{Not used yet}
}
\value{
New instance of rocker class
}
\description{
Function generates a new \link{R6} database handling interface with \link{DBI} backend.
For more information, see \link[rocker:rocker]{rocker} class description.
}
\examples{
db <- rocker::newDB()
}
\seealso{
Other rocker: 
\code{\link{rocker-R6-class}},
\code{\link{rocker-README}},
\code{\link{rocker-S3-functions}},
\code{\link{rocker-package}}
}
\concept{rocker}
