% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{removeTable}
\alias{removeTable}
\title{Remove table.}
\usage{
removeTable(db, name, ...)
}
\arguments{
\item{db}{rocker object}

\item{name}{Table name}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:dbRemoveTable]{DBI::dbRemoveTable()}}}
}
\value{
Invisible self
}
\description{
Remove table.
}
