% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{rollback}
\alias{rollback}
\title{Rollback transaction.}
\usage{
rollback(db, ...)
}
\arguments{
\item{db}{rocker object}

\item{...}{Optional, additional suitable parameters passed to \code{\link[DBI:transactions]{DBI::dbRollback()}}}
}
\value{
Invisible self
}
\description{
Rollback transaction.
}
